requirements.txt
streamlit
tensorflow
numpy
pillow
plotly
pandas
scikit-learn
plotly
matplotlib